+++ alinkjs/packages/assetlink-plugin-dev-support/src/index.js
@@ -11,15 +11,15 @@ const SERVE_DEV_HOST = process.env.ASSET_LINK_PLUGIN_SERVING_DEV_HOST || 'http:/
 /**
  * Libraries that are available for import in Asset Link plugins.
  * 
- * 'buffer/' 6.0.3 - https://www.npmjs.com/package/buffer
- * 'assetlink-plugin-api' - https://symbioquine.github.io/farmOS_asset_link/global.html
- * 'vue' 3.0.0 - https://vuejs.org/api/
- * 'vue-router' 4.0.0 - https://router.vuejs.org/api/
- * 'quasar' 2.6.0 - https://quasar.dev/docs
- * 'ngeohash' 0.6.3 - https://www.npmjs.com/package/ngeohash
- * 'haversine-distance' 1.2.1 - https://www.npmjs.com/package/haversine-distance
- * 'jmespath' 0.16.0 - https://www.npmjs.com/package/jmespath
- * 'micromustache' 8.0.3 - https://www.npmjs.com/package/micromustache
+ * - 'buffer/' 6.0.3 - https://www.npmjs.com/package/buffer
+ * - 'assetlink-plugin-api' - https://symbioquine.github.io/farmOS_asset_link/global.html
+ * - 'vue' 3.4.15 - https://vuejs.org/api/
+ * - 'vue-router' 4.0.0 - https://router.vuejs.org/api/
+ * - 'quasar' 2.6.0 - https://quasar.dev/docs
+ * - 'ngeohash' 0.6.3 - https://www.npmjs.com/package/ngeohash
+ * - 'haversine-distance' 1.2.1 - https://www.npmjs.com/package/haversine-distance
+ * - 'jmespath' 0.16.0 - https://www.npmjs.com/package/jmespath
+ * - 'micromustache' 8.0.3 - https://www.npmjs.com/package/micromustache
  * 
  * ### Usage in webpack.config.js
  *