+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Expose the farmOS instance name, version, and system of measurement
 
+### Fixed
+
+- Cleanup a few minor issues with the documentation
+
 ## [1.0.0-alpha13] - 2024-04-14
 
 ### Added